src/q11_serving_bert_model/fastapi/requirements.txt
fastapi
aiofiles
jinja2
python-multipart
tensorflow
bert-for-tf2
numpy